+++ op-service/rethdb-reader/.gitignore
@@ -1,2 +1,5 @@
 # Target
 target/
+
+# Bindings
+rdb.h
