+++ op-service/rethdb-reader/Cargo.toml
@@ -12,3 +12,4 @@ crate-type = ["cdylib"]
 reth = { git = "https://github.com/paradigmxyz/reth.git" }
 serde = "1.0.190"
 serde_json = "1.0.107"
+anyhow = "1.0.75"
